requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
aiohttp>=3.9.0
pyyaml>=6.0
inotify-simple>=1.3.5
psutil>=5.9.0
